# api-docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LandingResponse": {
        "properties": {
          "accent_color": {
            "title": "Accent Color",
            "type": "string"
          },
          "accent_palette": {
            "$ref": "#/components/schemas/AccentPaletteResponse"
          },
          "banners": {
            "items": {
              "$ref": "#/components/schemas/app__api__routers__public__BannerResponse"
            },
            "title": "Banners",
            "type": "array"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/ConversionBlockResponse"
            },
            "title": "Blocks",
            "type": "array"
          },
          "blocks_accent_color": {
            "title": "Blocks Accent Color",
            "type": "string"
          },
          "blocks_accent_palette": {
            "$ref": "#/components/schemas/AccentPaletteResponse"
          },
          "blocks_dark_mode": {
            "default": false,
            "title": "Blocks Dark Mode",
            "type": "boolean"
          },
          "cta_animation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cta Animation"
          },
          "cta_backgrounds": {
            "items": {
              "$ref": "#/components/schemas/CtaBackgroundResponse"
            },
            "title": "Cta Backgrounds",
            "type": "array"
          },
          "cta_band_style": {
            "title": "Cta Band Style",
            "type": "string"
          },
          "cta_color_modes": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Cta Color Modes",
            "type": "object"
          },
          "cta_positions": {
            "items": {
              "type": "integer"
            },
            "title": "Cta Positions",
            "type": "array"
          },
          "cta_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cta Text"
          },
          "cta_text_overrides": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Cta Text Overrides",
            "type": "object"
          },
          "default_offer_quantity": {
            "title": "Default Offer Quantity",
            "type": "integer"
          },
          "form_accent_color": {
            "title": "Form Accent Color",
            "type": "string"
          },
          "form_accent_palette": {
            "$ref": "#/components/schemas/AccentPaletteResponse"
          },
          "form_presentation": {
            "title": "Form Presentation",
            "type": "string"
          },
          "landing_id": {
            "title": "Landing Id",
            "type": "integer"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/app__api__routers__public__LandingOfferResponse"
            },
            "title": "Offers",
            "type": "array"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "product_price": {
            "title": "Product Price",
            "type": "number"
          },
          "product_sku": {
            "title": "Product Sku",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "variant_options": {
            "items": {
              "type": "object"
            },
            "title": "Variant Options",
            "type": "array"
          }
        },
        "required": [
          "landing_id",
          "product_id",
          "product_name",
          "product_sku",
          "product_price",
          "slug",
          "banners",
          "cta_positions",
          "cta_backgrounds",
          "form_presentation",
          "cta_band_style",
          "accent_color",
          "accent_palette",
          "form_accent_color",
          "form_accent_palette",
          "blocks_accent_color",
          "blocks_accent_palette",
          "offers",
          "default_offer_quantity",
          "blocks"
        ],
        "title": "LandingResponse",
        "type": "object"
      },
...
      "OrderCreateRequest": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "address1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address1"
          },
          "address2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address2"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "department": {
            "title": "Department",
            "type": "string"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "landing_slug": {
            "title": "Landing Slug",
            "type": "string"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "variant_selections": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Variant Selections",
            "type": "array"
          }
        },
        "required": [
          "landing_slug",
          "full_name",
          "phone",
          "department",
          "city",
          "address",
          "quantity"
        ],
        "title": "OrderCreateRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "The same session token supplied as `Authorization: Bearer <token>`, for clients that cannot hold cookies. Accepted everywhere the cookie is.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Session cookie issued by `POST /api/auth/login`. `HttpOnly`, `SameSite=Strict`, and `Secure` outside local development, so a browser sends it automatically and page scripts cannot read it. This is the transport browser clients should use.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "summary": "Cash-on-delivery commerce API for a single merchant: public landing delivery and checkout, plus the authenticated Administrator surface.",
    "title": "Easy Funnel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/public/landings/{slug}": {
      "get": {
        "description": "Return landing payload for public display.\n\nReturns identical 404 for unknown, draft, paused, or retired slugs\n(no information disclosure).",
        "operationId": "get_public_landing_api_public_landings__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Landing",
        "tags": [
          "public"
        ]
      }
    },
    "/api/public/landings/{slug}/cta-click": {
      "post": {
        "description": "Record a CTA click.",
        "operationId": "record_cta_click_api_public_landings__slug__cta_click_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Record Cta Click Api Public Landings  Slug  Cta Click Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Cta Click",
        "tags": [
          "public"
        ]
      }
    },
    "/api/public/landings/{slug}/view": {
      "post": {
        "description": "Record a landing view (no contact fields).",
        "operationId": "record_landing_view_api_public_landings__slug__view_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Record Landing View Api Public Landings  Slug  View Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Landing View",
        "tags": [
          "public"
        ]
      }
    },
...
    "/api/public/orders": {
      "post": {
        "description": "Submit a COD order with fraud evaluation.\n\nThe request IP address and user agent are taken from the trusted\nedge-forwarded request headers (Requirement 5.9-5.11), never from\nclient-supplied body or query values.\n\nReturns 201 on persisted result, 422 on validation error,\n503 on persistence failure (no partial order).",
        "operationId": "create_order_api_public_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Order",
        "tags": [
          "public"
        ]
      }
    },
...
  },
...
}